# BE/PHP/security/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/users:
    get:
      summary: ''
      operationId: getApiUsers
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  -
                    UserId: 1
                    FirstName: John
                    LastName: Doe
                    Email: john.doe@example.com
                    Password: password1
                    Avatar: avatar1.jpg
                    Address: '123 Main St'
                    PhoneNumber: '1234567890'
                    Gender: M
                    DoB: '1985-05-15 00:00:00.000'
                    Status: Active
                    RoleId: '1'
                    CityId: '1'
                  -
                    UserId: 2
                    FirstName: Jane
                    LastName: Smith
                    Email: jane.smith@example.com
                    Password: password2
                    Avatar: avatar2.jpg
                    Address: '456 Oak Ave'
                    PhoneNumber: '0987654321'
                    Gender: F
                    DoB: '1990-07-20 00:00:00.000'
                    Status: Active
                    RoleId: '2'
                    CityId: '2'
                  -
                    UserId: 3
                    FirstName: Michael
                    LastName: Johnson
                    Email: michael.johnson@example.com
                    Password: password3
                    Avatar: avatar3.jpg
                    Address: '789 Pine Rd'
                    PhoneNumber: '1122334455'
                    Gender: M
                    DoB: '1975-12-01 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '3'
                  -
                    UserId: 4
                    FirstName: Emily
                    LastName: Davis
                    Email: emily.davis@example.com
                    Password: password4
                    Avatar: avatar4.jpg
                    Address: '321 Maple St'
                    PhoneNumber: '2233445566'
                    Gender: F
                    DoB: '1988-09-30 00:00:00.000'
                    Status: Active
                    RoleId: '2'
                    CityId: '4'
                  -
                    UserId: 5
                    FirstName: Chris
                    LastName: Brown
                    Email: chris.brown@example.com
                    Password: password5
                    Avatar: avatar5.jpg
                    Address: '654 Birch Ln'
                    PhoneNumber: '3344556677'
                    Gender: M
                    DoB: '1992-11-10 00:00:00.000'
                    Status: Inactive
                    RoleId: '4'
                    CityId: '5'
                  -
                    UserId: 6
                    FirstName: Ashley
                    LastName: Wilson
                    Email: ashley.wilson@example.com
                    Password: password6
                    Avatar: avatar6.jpg
                    Address: '987 Cedar Dr'
                    PhoneNumber: '4455667788'
                    Gender: F
                    DoB: '1993-08-22 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '6'
                  -
                    UserId: 7
                    FirstName: David
                    LastName: Miller
                    Email: david.miller@example.com
                    Password: password7
                    Avatar: avatar7.jpg
                    Address: '123 Spruce Ct'
                    PhoneNumber: '5566778899'
                    Gender: M
                    DoB: '1981-02-14 00:00:00.000'
                    Status: Active
                    RoleId: '1'
                    CityId: '7'
                  -
                    UserId: 8
                    FirstName: Jessica
                    LastName: Martinez
                    Email: jessica.martinez@example.com
                    Password: password8
                    Avatar: avatar8.jpg
                    Address: '456 Elm Pl'
                    PhoneNumber: '6677889900'
                    Gender: F
                    DoB: '1987-03-25 00:00:00.000'
                    Status: Inactive
                    RoleId: '2'
                    CityId: '8'
                  -
                    UserId: 9
                    FirstName: Daniel
                    LastName: Garcia
                    Email: daniel.garcia@example.com
                    Password: password9
                    Avatar: avatar9.jpg
                    Address: '789 Fir St'
                    PhoneNumber: '7788990011'
                    Gender: M
                    DoB: '1995-10-12 00:00:00.000'
                    Status: Active
                    RoleId: '4'
                    CityId: '9'
                  -
                    UserId: 10
                    FirstName: Laura
                    LastName: Taylor
                    Email: laura.taylor@example.com
                    Password: password10
                    Avatar: avatar10.jpg
                    Address: '321 Oak Blvd'
                    PhoneNumber: '8899001122'
                    Gender: F
                    DoB: '1980-01-05 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '10'
                  -
                    UserId: 11
                    FirstName: Kevin
                    LastName: Anderson
                    Email: kevin.anderson@example.com
                    Password: password11
                    Avatar: avatar11.jpg
                    Address: '1111 King St'
                    PhoneNumber: '1231231234'
                    Gender: M
                    DoB: '1982-02-15 00:00:00.000'
                    Status: Active
                    RoleId: '2'
                    CityId: '11'
                  -
                    UserId: 12
                    FirstName: Sarah
                    LastName: Clark
                    Email: sarah.clark@example.com
                    Password: password12
                    Avatar: avatar12.jpg
                    Address: '2222 Queen Ave'
                    PhoneNumber: '2342342345'
                    Gender: F
                    DoB: '1994-03-18 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '12'
                  -
                    UserId: 13
                    FirstName: Robert
                    LastName: Lewis
                    Email: robert.lewis@example.com
                    Password: password13
                    Avatar: avatar13.jpg
                    Address: '3333 Prince Rd'
                    PhoneNumber: '3453453456'
                    Gender: M
                    DoB: '1989-04-20 00:00:00.000'
                    Status: Active
                    RoleId: '1'
                    CityId: '13'
                  -
                    UserId: 14
                    FirstName: Emily
                    LastName: Walker
                    Email: emily.walker@example.com
                    Password: password14
                    Avatar: avatar14.jpg
                    Address: '4444 Duke St'
                    PhoneNumber: '4564564567'
                    Gender: F
                    DoB: '1996-05-22 00:00:00.000'
                    Status: Inactive
                    RoleId: '2'
                    CityId: '14'
                  -
                    UserId: 15
                    FirstName: James
                    LastName: Hall
                    Email: james.hall@example.com
                    Password: password15
                    Avatar: avatar15.jpg
                    Address: '5555 Duchess Ave'
                    PhoneNumber: '5675675678'
                    Gender: M
                    DoB: '1983-06-25 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '15'
                  -
                    UserId: 16
                    FirstName: Olivia
                    LastName: Young
                    Email: olivia.young@example.com
                    Password: password16
                    Avatar: avatar16.jpg
                    Address: '6666 Earl Ct'
                    PhoneNumber: '6786786789'
                    Gender: F
                    DoB: '1988-07-28 00:00:00.000'
                    Status: Active
                    RoleId: '4'
                    CityId: '16'
                  -
                    UserId: 17
                    FirstName: William
                    LastName: King
                    Email: william.king@example.com
                    Password: password17
                    Avatar: avatar17.jpg
                    Address: '7777 Count Pl'
                    PhoneNumber: '7897897890'
                    Gender: M
                    DoB: '1987-08-30 00:00:00.000'
                    Status: Inactive
                    RoleId: '2'
                    CityId: '17'
                  -
                    UserId: 18
                    FirstName: Ava
                    LastName: Scott
                    Email: ava.scott@example.com
                    Password: password18
                    Avatar: avatar18.jpg
                    Address: '8888 Baron Dr'
                    PhoneNumber: '8908908901'
                    Gender: F
                    DoB: '1991-09-02 00:00:00.000'
                    Status: Active
                    RoleId: '1'
                    CityId: '18'
                  -
                    UserId: 19
                    FirstName: Henry
                    LastName: Green
                    Email: henry.green@example.com
                    Password: password19
                    Avatar: avatar19.jpg
                    Address: '9999 Viscount St'
                    PhoneNumber: '9019019012'
                    Gender: M
                    DoB: '1985-10-05 00:00:00.000'
                    Status: Active
                    RoleId: '4'
                    CityId: '19'
                  -
                    UserId: 20
                    FirstName: Sophia
                    LastName: Adams
                    Email: sophia.adams@example.com
                    Password: password20
                    Avatar: avatar20.jpg
                    Address: '1010 Marquis Ln'
                    PhoneNumber: '0120120123'
                    Gender: F
                    DoB: '1990-11-08 00:00:00.000'
                    Status: Active
                    RoleId: '3'
                    CityId: '20'
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiUsers
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                FirstName:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: oainiuuovrdwztvpje
                LastName:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: dghvnzhcq
                Email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 50 characters.'
                  example: paucek.lura@example.net
                Password:
                  type: string
                  description: 'Must be at least 6 characters. Must not be greater than 12 characters.'
                  example: 'uX`pBR|]"A]8p'
                Avatar:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: ytrjpqtgaugauom
                Address:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: nkw
                PhoneNumber:
                  type: string
                  description: 'Must be 10 characters.'
                  example: gtfebzkxjm
                Gender:
                  type: string
                  description: ''
                  example: O
                  enum:
                    - M
                    - F
                    - O
                DoB:
                  type: string
                  description: 'Must be a valid date. Must be a date before <code>today</code>.'
                  example: '2007-09-03'
                Status:
                  type: string
                  description: ''
                  example: Active
                  enum:
                    - Active
                    - Inactive
                RoleId:
                  type: integer
                  description: ''
                  example: 5
                CityId:
                  type: integer
                  description: ''
                  example: 15
              required:
                - FirstName
                - LastName
                - Email
                - Password
                - PhoneNumber
                - DoB
                - RoleId
                - CityId
      security: []
  '/api/users/{UserId}':
    get:
      summary: ''
      operationId: getApiUsersUserId
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  UserId: 1
                  FirstName: John
                  LastName: Doe
                  Email: john.doe@example.com
                  Password: password1
                  Avatar: avatar1.jpg
                  Address: '123 Main St'
                  PhoneNumber: '1234567890'
                  Gender: M
                  DoB: '1985-05-15 00:00:00.000'
                  Status: Active
                  RoleId: '1'
                  CityId: '1'
                properties:
                  UserId:
                    type: integer
                    example: 1
                  FirstName:
                    type: string
                    example: John
                  LastName:
                    type: string
                    example: Doe
                  Email:
                    type: string
                    example: john.doe@example.com
                  Password:
                    type: string
                    example: password1
                  Avatar:
                    type: string
                    example: avatar1.jpg
                  Address:
                    type: string
                    example: '123 Main St'
                  PhoneNumber:
                    type: string
                    example: '1234567890'
                  Gender:
                    type: string
                    example: M
                  DoB:
                    type: string
                    example: '1985-05-15 00:00:00.000'
                  Status:
                    type: string
                    example: Active
                  RoleId:
                    type: string
                    example: '1'
                  CityId:
                    type: string
                    example: '1'
      tags:
        - Endpoints
      security: []
    put:
      summary: ''
      operationId: putApiUsersUserId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                FirstName:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: mvlzlgucvfphvlkw
                LastName:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: apkufcf
                Email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 50 characters.'
                  example: ondricka.gust@example.net
                Password:
                  type: string
                  description: 'Must be at least 6 characters. Must not be greater than 12 characters.'
                  example: e0oo-4
                Avatar:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: xhvkfkhuwpzrncwjs
                Address:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: zofmjazlvxhjnebkztaslvilv
                PhoneNumber:
                  type: string
                  description: 'Must be 10 characters.'
                  example: smyafyrbpz
                Gender:
                  type: string
                  description: ''
                  example: M
                  enum:
                    - M
                    - F
                    - O
                DoB:
                  type: string
                  description: 'Must be a valid date. Must be a date before <code>today</code>.'
                  example: '2017-04-28'
                Status:
                  type: string
                  description: ''
                  example: Inactive
                  enum:
                    - Active
                    - Inactive
                RoleId:
                  type: integer
                  description: ''
                  example: 2
                CityId:
                  type: integer
                  description: ''
                  example: 5
              required:
                - FirstName
                - LastName
                - Email
                - Password
                - PhoneNumber
                - DoB
                - RoleId
                - CityId
      security: []
    delete:
      summary: ''
      operationId: deleteApiUsersUserId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: UserId
        description: ''
        example: 1
        required: true
        schema:
          type: integer
...